@@ -9,12 +9,13 @@
     "bootstrap": "^3.2.0",
     "director": "^1.2.3",
     "flux": "^2.0.1",
-    "react": "^0.11.2"
+    "lodash": "^2.4.1",
+    "react": "^0.12.0"
   },
   "devDependencies": {
-    "browser-sync": "^1.5.8",
+    "browser-sync": "^1.6.1",
     "del": "^0.1.3",
-    "gulp": "^3.8.8",
+    "gulp": "^3.8.9",
     "gulp-autoprefixer": "^1.0.1",
     "gulp-cache": "^0.2.4",
     "gulp-changed": "^1.0.0",
@@ -22,31 +23,31 @@
     "gulp-gh-pages": "^0.4.0",
     "gulp-htmlmin": "^0.2.0",
     "gulp-if": "^1.2.5",
-    "gulp-imagemin": "^1.1.0",
+    "gulp-imagemin": "^1.2.1",
     "gulp-jsbeautifier": "^0.0.3",
-    "gulp-jshint": "^1.8.5",
+    "gulp-jshint": "^1.8.6",
     "gulp-less": "^1.3.6",
     "gulp-load-plugins": "^0.7.0",
     "gulp-minify-css": "^0.3.11",
     "gulp-plumber": "^0.6.6",
-    "gulp-render": "^0.0.4",
+    "gulp-render": "^0.1.1",
     "gulp-replace": "^0.4.0",
     "gulp-size": "^1.1.0",
     "gulp-uglify": "^1.0.1",
     "gulp-util": "^3.0.1",
     "jest-cli": "^0.1.18",
-    "jshint": "^2.5.6",
+    "jshint": "^2.5.7",
     "jshint-loader": "^0.8.0",
     "jshint-stylish": "^1.0.0",
-    "jsx-loader": "^0.11.2",
+    "jsx-loader": "^0.12.0",
     "merge-stream": "^0.1.6",
     "minimist": "^1.1.0",
     "protractor": "^1.3.1",
     "psi": "^0.1.4",
-    "react-tools": "^0.11.2",
+    "react": "^0.12.0",
     "run-sequence": "^1.0.1",
     "url-loader": "^0.5.5",
-    "webpack": "^1.4.8",
+    "webpack": "^1.4.9",
     "webpack-dev-server": "^1.6.5"
   },
   "jest": {
