@@ -1,7 +1,3 @@
-/**
- * @jsx React.DOM
- */
-
 'use strict';
 
 var React = require('react');
@@ -10,7 +6,7 @@ var Link = require('./Link.jsx');
 var Navbar = React.createClass({
   render() {
     return (
-      <div className="navbar-top">
+      <div className="navbar-top" role="navigation">
         <div className="container">
           <Link className="navbar-brand row" to="/">
             <img src="/images/logo-small.png" width="38" height="38" alt="React" />
